Changed discord_id field from int to long.

+++ webservice/src/main/java/com/tategreeves/webservice/Repository/AuthenticationRepository.java
@@ -3,5 +3,5 @@ package com.tategreeves.webservice.Repository;
 import com.tategreeves.webservice.Model.Token;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface AuthenticationRepository extends MongoRepository<Token, Integer> {
+public interface AuthenticationRepository extends MongoRepository<Token, Long> {
 }
